hotel_veranum/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
marshmallow==3.17.0
packaging==24.1
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.0
transbank-sdk==5.0.0
tzdata==2024.1
urllib3==2.2.2
